+++ packages/router/types/types/curi.d.ts
@@ -42,6 +42,7 @@ export interface NavigationDetails {
     cancelled?: () => void;
     finished?: () => void;
 }
+export declare type CancelNavigateCallbacks = () => void;
 export interface CuriRouter {
     refresh: (routeArray?: UserRoutes) => void;
     observe: (fn: Observer, options?: ResponseHandlerOptions) => RemoveObserver;
@@ -50,5 +51,5 @@ export interface CuriRouter {
     route: Interactions;
     history: History;
     current(): CurrentResponse;
-    navigate(options: NavigationDetails): void;
+    navigate(options: NavigationDetails): CancelNavigateCallbacks;
 }